Add PyGithub to requirements, move tasks fixtures out
This is to prepare for new Github-related functionality and tests
