+++ tests/account/submission_test.py
@@ -15,6 +15,7 @@ from ocflib.account.submission import username_pending
 from tests.account.creation_test import fake_credentials  # noqa
 from tests.account.creation_test import fake_new_account_request  # noqa
 from tests.account.creation_test import mock_rsa_key  # noqa
+from tests.fixtures_test import celery_app  # noqa
 
 
 @pytest.fixture
@@ -60,36 +61,6 @@ class TestUserHasRequestPending:
         session.add(StoredNewAccountRequest.from_request(fake_new_account_request, 'reason'))
         session.commit()
         assert not user_has_request_pending(session, fake_new_account_request)
-
-
-@pytest.fixture
-def celery_app():
-    sent_messages = []
-
-    def mock_celery_task(f):
-        def update_state(**kwargs):
-            pass
-
-        # wrap everything in Mock so we can track calls
-        return mock.Mock(
-            side_effect=f,
-            delay=mock.Mock(),
-            update_state=mock.Mock(side_effect=update_state),
-        )
-
-    @contextmanager
-    def dispatcher():
-        def send(**kwargs):
-            sent_messages.append(kwargs)
-        yield mock.Mock(send=send)
-
-    return mock.Mock(
-        task=mock_celery_task,
-        _sent_messages=sent_messages,
-        **{
-            'events.default_dispatcher': dispatcher,
-        }
-    )
 
 
 @pytest.yield_fixture
